+++ backend/src/test/java/com/onair/hearit/auth/application/AuthServiceTest.java
@@ -9,8 +9,8 @@ import com.onair.hearit.auth.dto.request.SignupRequest;
 import com.onair.hearit.auth.dto.response.LoginTokenResponse;
 import com.onair.hearit.auth.infrastructure.client.KakaoUserInfoClient;
 import com.onair.hearit.auth.infrastructure.jwt.JwtTokenProvider;
-import com.onair.hearit.auth.infrastructure.jwt.RefreshToken;
-import com.onair.hearit.auth.infrastructure.jwt.RefreshTokenRepository;
+import com.onair.hearit.auth.domain.RefreshToken;
+import com.onair.hearit.auth.infrastructure.repository.RefreshTokenRepository;
 import com.onair.hearit.common.exception.custom.InvalidInputException;
 import com.onair.hearit.common.exception.custom.UnauthorizedException;
 import com.onair.hearit.config.TestJpaAuditingConfig;
