+++ backend/src/test/java/com/onair/hearit/auth/presentation/AuthControllerTest.java
@@ -10,8 +10,8 @@ import com.onair.hearit.auth.dto.request.TokenReissueRequest;
 import com.onair.hearit.auth.dto.response.LoginTokenResponse;
 import com.onair.hearit.auth.dto.response.TokenReissueResponse;
 import com.onair.hearit.auth.infrastructure.jwt.JwtTokenProvider;
-import com.onair.hearit.auth.infrastructure.jwt.RefreshToken;
-import com.onair.hearit.auth.infrastructure.jwt.RefreshTokenRepository;
+import com.onair.hearit.auth.domain.RefreshToken;
+import com.onair.hearit.auth.infrastructure.repository.RefreshTokenRepository;
 import com.onair.hearit.domain.Member;
 import com.onair.hearit.fixture.DbHelper;
 import com.onair.hearit.fixture.IntegrationTest;
